exercise 6

+++ src/java/simpledb/storage/HeapPage.java
@@ -95,7 +95,7 @@ public class HeapPage implements Page {
     private int getHeaderSize() {
 
         // TODO: some code goes here
-        return (int)Math.ceil(getNumTuples() / 8);
+        return (int)Math.ceil((double)getNumTuples() / 8);
 
     }
